seperated seed and db model to two separate files and successful api call between client, server, db
