+++ database/seed.js
@@ -1,15 +1,7 @@
 /* eslint-disable no-plusplus */
 const mongoose = require('mongoose');
 const faker = require('faker');
-
-mongoose.connect('mongodb://localhost/house');
-
-const houseSchema = new mongoose.Schema({
-  house_id: Number,
-  pictures: [{ description: String, url: String, isVerified: Boolean }],
-});
-
-const House = mongoose.model('House', houseSchema);
+const House = require('./db.js');
 
 mongoose.connection.on('error', console.error.bind(console, 'connection error:'));
 mongoose.connection.once('open', () => {
@@ -49,6 +41,3 @@ mongoose.connection.once('open', () => {
     })
     .catch(err => console.log(err));
 });
-
-module.exports = House; 
-
