+++ package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "react-dev": "webpack -d --watch",
     "server-dev": "nodemon server/index.js",
-    "seed": "node database/index.js"
+    "seed": "node database/seed.js"
   },
   "keywords": [],
   "author": "",
