+++ server/index.js
@@ -1,15 +1,14 @@
 const express = require('express');
 const mongoose = require('mongoose');
-const House = require('../database/index.js');
+const House = require('../database/db.js');
 
 const app = express();
 const port = 3002;
 
 app.use(express.static(`${__dirname}/../client/dist`));
 
-mongoose.connect('mongodb://localhost/house');
-
 app.get('/house/:houseId', (req, res) => {
+  console.log(req.params)
   House.findOne({ house_id: req.params.houseId }, (err, house) => {
     if (err) {
       res.status(404);
